+++ scripts/tests_embedded_scripts/test_memfault_gdb.py
@@ -22,7 +22,7 @@ scripts_dir = os.path.dirname(tests_dir)
 sys.path.insert(0, scripts_dir)
 
 
-from memfault_gdb import (  # noqa: E402 M900
+from memfault_gdb import (  # noqa: E402
     ArmCortexMCoredumpArch,
     MemfaultConfig,
     MemfaultCoredump,
@@ -221,7 +221,6 @@ def _settings_coredump_allowed(test_config):
 
 
 def test_parse_maintenance_info_sections_no_file():
-
     fn, sections = parse_maintenance_info_sections(
         """Remote serial target in gdb-specific protocol:
 Debugging a target over a serial line.
@@ -232,7 +231,6 @@ Debugging a target over a serial line.
 
 
 def test_parse_maintenance_info_sections_with_file(maintenance_info_sections_fixture, fake_elf):
-
     fn, sections = parse_maintenance_info_sections(maintenance_info_sections_fixture)
     assert fn == str(fake_elf)
     assert len(sections) == 35
@@ -255,7 +253,6 @@ def test_read_current_registers(mocker, info_reg_all_fixture):
 
 
 def test_should_capture_section():
-
     # Never capture .text, even when NOT marked READONLY:
     assert not should_capture_section(Section(0, 10, ".text", read_only=False))
 
@@ -273,7 +270,6 @@ def test_should_capture_section():
 
 
 def test_armv7_get_used_ram_base_addresses():
-
     sections = (
         Section(0x30000000, 10, "", read_only=True),
         Section(0x70000000, 10, "", read_only=True),
@@ -291,7 +287,6 @@ def test_armv7_get_used_ram_base_addresses():
 
 
 def test_read_memory_until_error_no_error():
-
     read_size = 1024
     size = 1024 * 1024
     inferior = MagicMock()
@@ -301,7 +296,6 @@ def test_read_memory_until_error_no_error():
 
 
 def test_read_memory_until_error_after_10k():
-
     read_size = 1024
     size = 1024 * 1024
     inferior = MagicMock()
@@ -320,12 +314,14 @@ def test_read_memory_until_error_after_10k():
 
 
 def test_coredump_writer(snapshot):
-
     arch = ArmCortexMCoredumpArch()
-    cd_writer = MemfaultCoredumpWriter(arch)
-    cd_writer.device_serial = "device_serial"
-    cd_writer.firmware_version = "1.2.3"
-    cd_writer.hardware_revision = "gdb-proto"
+    device_serial = "device_serial"
+    software_type = "main"
+    software_version = "1.2.3"
+    hardware_revision = "gdb-proto"
+    cd_writer = MemfaultCoredumpWriter(
+        arch, device_serial, software_type, software_version, hardware_revision
+    )
     cd_writer.trace_reason = 5
     cd_writer.regs = [
         {
@@ -363,7 +359,6 @@ def test_coredump_writer(snapshot):
 
 
 def test_http_basic_auth():
-
     headers = add_basic_auth("martijn@memfault.com", "open_sesame", {"Foo": "Bar"})
     assert headers == {
         "Foo": "Bar",
@@ -379,7 +374,7 @@ def _gdb_for_coredump(mocker, maintenance_info_sections_fixture, _settings_cored
     def _gdb_execute(cmd, to_string):
         if cmd == "maintenance info sections":
             return maintenance_info_sections_fixture
-        if cmd == "info reg all":
+        if cmd == "info all-registers":
             return "r0\t0\n"
         if cmd == "info threads":
             return ""
@@ -465,6 +460,65 @@ def test_coredump_command_not_allowing(
 
     stdout = capsys.readouterr().out
     assert "Aborting" in stdout
+
+
+@pytest.mark.usefixtures("_gdb_for_coredump")
+def test_coredump_all_overrides(http_expect_request, test_config, mocker):
+    """Test coredump command with all overrides set"""
+
+    # Verify coredump is uploaded
+    test_config.ingress_uri = "https://custom-ingress.memfault.com"
+    http_expect_request(
+        "https://custom-ingress.memfault.com/api/v0/upload/coredump",
+        "POST",
+        ANY,
+        {"Content-Type": "application/octet-stream", "Memfault-Project-Key": TEST_PROJECT_KEY},
+        200,
+        {},
+    )
+
+    writer = MagicMock()
+    mocker.patch("memfault_gdb.MemfaultCoredumpWriter", writer)
+
+    hardware_revision = "TESTREVISION"
+    software_version = "TESTVERSION"
+    software_type = "TESTTYPE"
+    device_serial = "TESTSERIAL"
+
+    cmd = MemfaultCoredump()
+    cmd.invoke(
+        "--project-key {} --hardware-revision {} --software-version {} --software-type {} --device-serial {}".format(
+            TEST_PROJECT_KEY, hardware_revision, software_version, software_type, device_serial
+        ),
+        True,
+    )
+
+    # Verify that all args were successfully extracted and passed to writer
+    writer.assert_called_once_with(
+        ANY, device_serial, software_type, software_version, hardware_revision
+    )
+
+
+@pytest.mark.parametrize(
+    ("cmd_option", "cmd_type"),
+    [
+        ("--hardware-revision", "hardware revision"),
+        ("--software-version", "software version"),
+        ("--software-type", "software type"),
+        ("--device-serial", "device serial"),
+    ],
+)
+def test_coredump_override_invalid_input(capsys, cmd_option, cmd_type):
+    """Test coredump command with invalid input for each override"""
+
+    invalid_input = 'inv"lid'
+
+    cmd = MemfaultCoredump()
+    cmd.invoke("--project-key {} {} {}".format(TEST_PROJECT_KEY, cmd_option, invalid_input), True)
+
+    # Verify an error is thrown for the invalid input
+    stderr = capsys.readouterr().err
+    assert "Invalid characters in {}: {}".format(cmd_type, invalid_input) in stderr
 
 
 @pytest.mark.parametrize(
@@ -576,7 +630,6 @@ https://app.memfault.com/organizations/acme-inc/projects/smart-sink/issues?live"
 
 
 def test_login_command_simple(http_expect_request, test_config):
-
     http_expect_request(
         "https://api.memfault.com/auth/me", "GET", None, TEST_AUTH_HEADERS, 200, {"id": 123}
     )
@@ -592,7 +645,6 @@ def test_login_command_simple(http_expect_request, test_config):
 
 
 def test_login_command_with_all_options(http_expect_request, test_config):
-
     test_api_uri = "http://dev-api.memfault.com:8000"
     test_ingress_uri = "http://dev-ingress.memfault.com"
 
@@ -640,7 +692,6 @@ def test_login_command_with_all_options(http_expect_request, test_config):
 def test_has_uploaded_symbols(
     http_expect_request, expected_result, resp_status, resp_body, test_config_with_login
 ):
-
     test_software_type = "main"
     test_software_version = "1.0.0"
     http_expect_request(
@@ -664,7 +715,6 @@ def test_has_uploaded_symbols(
 
 
 def test_post_chunk_data(http_expect_request):
-
     base_uri = "https://example.chunks.memfault.com"
     chunk_data = bytearray([1, 2, 3, 4])
     device_serial = "GDB_TESTSERIAL"